fix(matechat): 删除多余的样式，精简代码
